Update to latest sources and hardware rendering

+++ server/sv_null.c
@@ -1,5 +1,6 @@
 // sv_null.c -- this file can stub out the entire server system
 // for pure net-only clients
+#include "../client/client.h"
 
 void SV_Init (void)
 {
@@ -9,7 +10,7 @@ void SV_Shutdown (char *finalmsg, qboolean reconnect)
 {
 }
 
-void SV_Frame (float time)
+void SV_Frame (double time)
 {
 }
